requirements.txt
streamlit
pinecone-client
groq
langchain-core
langchain-text-splitters
playwright
python-dotenv
requests
beautifulsoup4
tqdm
